make readme simpler
