+++ README.md
@@ -17,8 +17,7 @@ Then you need to add right after `require 'bundler/setup'`:
 
 ```ruby
 require 'bundler/setup'
-require 'bootscale'
-Bootscale.setup(cache_directory: 'tmp/bootscale')
+require 'bootscale/setup'
 ```
 
 If your application is a Rails application, you will find this in `config/boot.rb`.
